+++ .gitignore
@@ -128,4 +128,5 @@ dmypy.json
 # Pyre type checker
 .pyre/
 
-FTV/analys_private_sprid_ej.py
+Data/
+Modaliteter/FTV/analys_private_sprid_ej.py
