+++ .vscode/settings.json
@@ -1,0 +1,20 @@
+{
+    "editor.rulers": [
+        88
+    ],
+    "editor.minimap.enabled": false,
+    "editor.fontSize": 12,
+    "editor.fontFamily": "Fira Code",
+    "editor.fontLigatures": true,
+   "python.linting.pylamaEnabled": true,
+    "terminal.integrated.fontSize": 12,
+    "files.exclude": {
+        "**/__pycache__": true
+    },
+    "python.languageServer": "Pylance",
+    "window.zoomLevel": -1,
+    "workbench.editorAssociations": {
+        "*.ipynb": "jupyter-notebook"
+    },
+    "terminal.integrated.defaultProfile.windows": "Command Prompt",
+}
